+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes 12.8
 ------------
 - Fix reassign existing Barcode to student Food Service account in Food_Service/Students/Accounts.php
+- Add `_skipDie()` function, used when skipping CSP report in SaveReport.php
 
 Changes in 12.7
 ---------------